new color and better sync
