+++ PomodoroExtension/Widget/Views/WidgetTaskRow.swift
@@ -36,24 +36,4 @@ struct WidgetTaskRow: View {
     }
 }
 
-extension ToDoStore.TaskFilter {
-    var systemImage: String {
-        switch self {
-        case .all: return "tray.full"
-        case .today: return "calendar.circle"
-        case .tomorrow: return "calendar.badge.plus"
-        case .thisWeek: return "calendar"
-        case .overdue: return "exclamationmark.triangle"
-        }
-    }
-    
-    var color: Color {
-        switch self {
-        case .all: return .gray
-        case .today: return .blue
-        case .tomorrow: return .green
-        case .thisWeek: return .purple
-        case .overdue: return .red
-        }
-    }
-}
+
